Move WanSyncStats and ConsistencyCheckResult to private API
